requirements.txt
click==8.1.8
lxml==5.3.0
lxml-stubs==0.5.1
python-dotenv==1.0.1
PyYAML==6.0.2
types-PyYAML==6.0.12.20241230
docstring-parser==0.16
litellm==1.73.6
geopandas==1.1.1
fastapi
uvicorn
python-multipart
openai
pydub
